backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
dj-database-url==2.1.0
psycopg[binary]==3.1.18
ruff==0.1.7
black==24.4.2
pytest==8.2.2
pytest-django==4.8.0
djangorestframework-simplejwt==5.3.1
